+++ core-ng/src/main/java/core/framework/impl/web/session/SessionStore.java
@@ -2,6 +2,7 @@ package core.framework.impl.web.session;
 
 import java.time.Duration;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * @author neo
@@ -9,7 +10,7 @@ import java.util.Map;
 interface SessionStore {
     Map<String, String> getAndRefresh(String sessionId, Duration sessionTimeout);
 
-    void save(String sessionId, Map<String, String> sessionData, Duration sessionTimeout);
+    void save(String sessionId, Map<String, String> values, Set<String> changedFields, Duration sessionTimeout);
 
     void invalidate(String sessionId);
 }