+++ core-ng/src/main/java/core/framework/impl/log/TraceLogger.java
@@ -71,7 +71,7 @@ public final class TraceLogger {
         }
     }
 
-    Writer createWriter(ActionLog log) {
+    private Writer createWriter(ActionLog log) {
         if (traceLogPath != null) {
             try {
                 String logPath = traceLogFilePath(traceLogPath.toString(), LocalDateTime.ofInstant(log.date, ZoneId.systemDefault()), log.action, log.id);