+++ core-ng/src/main/java/core/framework/impl/template/fragment/Fragment.java
@@ -5,6 +5,6 @@ import core.framework.impl.template.TemplateContext;
 /**
  * @author neo
  */
-public interface Fragment {
+interface Fragment {
     void process(StringBuilder builder, TemplateContext context);
 }