+++ core-ng/src/main/java/core/framework/impl/web/request/RequestImpl.java
@@ -5,7 +5,6 @@ import core.framework.api.http.HTTPMethod;
 import core.framework.api.util.Encodings;
 import core.framework.api.util.Exceptions;
 import core.framework.api.util.Maps;
-import core.framework.api.util.Strings;
 import core.framework.api.web.CookieSpec;
 import core.framework.api.web.MultipartFile;
 import core.framework.api.web.Request;
@@ -15,12 +14,9 @@ import core.framework.impl.json.JSONMapper;
 import core.framework.impl.web.BeanValidator;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.server.handlers.Cookie;
-import io.undertow.server.handlers.form.FormData;
-import io.undertow.util.Headers;
 
 import java.io.UncheckedIOException;
 import java.lang.reflect.Type;
-import java.util.Deque;
 import java.util.Map;
 import java.util.Optional;
 
@@ -29,6 +25,11 @@ import java.util.Optional;
  */
 public final class RequestImpl implements Request {
     public final PathParams pathParams = new PathParams();
+
+    final Map<String, String> queryParams = Maps.newHashMap();
+    final Map<String, String> formParams = Maps.newHashMap();
+    final Map<String, MultipartFile> files = Maps.newHashMap();
+
     private final HttpServerExchange exchange;
     private final BeanValidator validator;
     public Session session;
@@ -38,7 +39,6 @@ public final class RequestImpl implements Request {
     int port;
     String requestURL;
     ContentType contentType;
-    FormData formData;
     byte[] body;
 
     public RequestImpl(HttpServerExchange exchange, BeanValidator validator) {
@@ -107,66 +107,39 @@ public final class RequestImpl implements Request {
 
     @Override
     public <T> Optional<T> queryParam(String name, Class<T> valueClass) {
-        Deque<String> values = exchange.getQueryParameters().get(name);
-        if (values == null) return Optional.empty();
-        String first = values.getFirst();
-        return Optional.of(URLParamParser.parse(first, valueClass));
+        String value = queryParams.get(name);
+        if (value == null) return Optional.empty();
+        return Optional.of(URLParamParser.parse(value, valueClass));
     }
 
     @Override
     public Map<String, String> queryParams() {
-        Map<String, String> params = Maps.newHashMap();
-        exchange.getQueryParameters().forEach((name, values) -> params.put(name, values.getFirst()));
-        return params;
+        return queryParams;
     }
 
     @Override
     public Optional<String> formParam(String name) {
-        FormData.FormValue value = formData().getFirst(name);
-        if (value == null) return Optional.empty();
-        return Optional.ofNullable(value.getValue());
+        return Optional.ofNullable(formParams.get(name));
     }
 
     @Override
     public Map<String, String> formParams() {
-        FormData formData = formData();
-        Map<String, String> form = Maps.newHashMap();
-        for (String name : formData) {
-            FormData.FormValue value = formData.getFirst(name);
-            if (!value.isFile()) form.put(name, value.getValue());
-        }
-        return form;
+        return formParams;
     }
 
     @Override
     public Optional<MultipartFile> file(String name) {
-        FormData.FormValue value = formData().getFirst(name);
-        if (value == null) return Optional.empty();
-        if (!value.isFile()) throw new BadRequestException("form body must be multipart, method=" + method + ", contentType=" + contentType);
-        if (Strings.isEmpty(value.getFileName())) return Optional.empty();  // browser passes empty file name if not choose file in form
-        return Optional.of(new MultipartFile(value.getPath(), value.getFileName(), value.getHeaders().getFirst(Headers.CONTENT_TYPE)));
+        return Optional.ofNullable(files.get(name));
     }
 
     @Override
     public Map<String, MultipartFile> files() {
-        FormData formData = formData();
-        Map<String, MultipartFile> files = Maps.newHashMap();
-        for (String name : formData) {
-            FormData.FormValue value = formData.getFirst(name);
-            if (value.isFile() && !Strings.isEmpty(value.getFileName()))
-                files.put(name, new MultipartFile(value.getPath(), value.getFileName(), value.getHeaders().getFirst(Headers.CONTENT_TYPE)));
-        }
         return files;
     }
 
     @Override
     public Optional<byte[]> body() {
         return Optional.ofNullable(body);
-    }
-
-    private FormData formData() {
-        if (formData == null) throw new BadRequestException("form body is required, method=" + method + ", contentType=" + contentType);
-        return formData;
     }
 
     @Override
@@ -181,12 +154,10 @@ public final class RequestImpl implements Request {
 
     private <T> T parseBean(Type instanceType) {
         if (method == HTTPMethod.GET || method == HTTPMethod.DELETE) {
-            Map<String, String> params = Maps.newHashMap();
-            exchange.getQueryParameters().forEach((name, values) -> params.put(name, values.element()));
-            return JSONMapper.fromMapValue(instanceType, params);
+            return JSONMapper.fromMapValue(instanceType, queryParams);
         } else if (method == HTTPMethod.POST || method == HTTPMethod.PUT) {
-            if (formData != null) {
-                return JSONMapper.fromMapValue(instanceType, formParams());
+            if (!formParams.isEmpty()) {
+                return JSONMapper.fromMapValue(instanceType, formParams);
             } else if (body != null && contentType != null && ContentType.APPLICATION_JSON.mediaType().equals(contentType.mediaType())) {
                 return JSONMapper.fromJSON(instanceType, body);
             }